+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Changelog
 
-## 1.8.0 (unreleased)
+## 1.8.0
 
 - Refactor SDK: `com.powersync:powersync-core` has an identical API, but now depends on 
   `com.powersync:powersync-common` where most logic is implemented.
   - __POTENTIALLY BREAKING CHANGE__: If you were injecting a `DatabaseDriverFactory` into Koin or Dagger, note that the
     `PowerSyncDatabase()` factory method now takes a more generic `PersistentConnectionFactory`.
   - If you're using `PowerSyncDatabase.inMemory`, you explicitly have to import `com.powersync.inMemory` now.
+- Update the PowerSync core extension to version 0.4.8.
+- Add the `soft` flag to `disconnectAndClear()` which keeps an internal copy of synced data in the database, allowing
+  faster re-sync if a compatible token is used in the next `connect()` call.
+- Add the `clear` parameter to `RawTable` to run a statement helping the core extension clear raw tables.
 
 ## 1.7.0